requirements.txt
azure-storage-blob==12.25.1
azure-ai-documentintelligence==1.0.2
azure-identity==1.23.0
python-dotenv==1.1.1
numpy==2.3.1
uvicorn==0.35.0
fastapi==0.116.1